% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrametoXML.R
\name{DataFrametoXML}
\alias{DataFrametoXML}
\title{Create an ED2 XML file with PFT parameters from an R dataframe}
\usage{
DataFrametoXML(df, outpath)
}
\arguments{
\item{df}{Full path to the dataframe file (character string) OR an R dataframe object}

\item{outpath}{Where the output should be saved (full path, character string)}
}
\value{
ED2-compatible XML file with PFT parameters, saved in outpath
}
\description{
Create an ED2 XML file with PFT parameters from an R dataframe
}
\examples{

df<-XMLtoDataFrame(xml="/mnt/odyssey/moorcroftfs5/mjohnston/gitruns/run005/Tonzi.xml")
outpath="/home/miriam/Desktop/DFtoXMLtest.xml"
DataFrametoXML(df,outpath)

}
