% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED2Summary.R
\name{ED2Summary}
\alias{ED2Summary}
\title{ED2 Model Result Diagnosis}
\usage{
ED2Summary(ED2INdir, analydir)
}
\arguments{
\item{ED2INdir}{Directory containing ED2IN (character string, with terminal "/")}

\item{analydir}{Directory containing the result h5 files (character string, usually ending in "analy", with terminal "/")}
}
\value{
Summary PDF document - an ED2 "sanity check" summarizing the plants and carbon/water/energy fluxes.
}
\description{
ED2 sanity check
}
\section{Important notes on this function}{
 \itemize{
\item ***This function will close & re-open your RStudio, deleting environment variables***
\item This function always saves the PDF as: \cr
ED2INdir/ED2Summary.pdf
\item This function has quite a few companion scripts, all in: \cr
ED2ToolsMRJ/ED2Summary_CompanionFxns
}
}

\examples{
ED2INdir<-"/mnt/odyssey/moorcroftfs5/mjohnston/ED2_Ashehad/Case7.6/run048/"
analydir<-"/mnt/odyssey/moorcroftfs5/mjohnston/ED2_Ashehad/Case7.6/run048/analy/"
ED2Summary(ED2INdir,analydir)

}
