% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractChronDates.R
\name{ExtractChronDates}
\alias{ExtractChronDates}
\title{Extract Chron Dates}
\usage{
ExtractChronDates(dir, tscale, perfile, first, last)
}
\arguments{
\item{dir}{Directory in which to search for ED2 output files (character string)}

\item{tscale}{Time scale for relevant files (single character; common options: "D","I")}

\item{perfile}{Number of times desired per file (optional, numeric, use only sometimes for I tscale, default =1)}

\item{first}{First file from which to extract (optional, character string, use with 'last' to extract between a set of files)}

\item{last}{Last file from which to extract (optional, character string, use with 'first' to extract between a set of files)}
}
\value{
'chron' format dates
}
\description{
Used to extract 'chron' dates from filenames in a directory
}
\examples{

dir="/mnt/odyssey/moorcroftfs5/mjohnston/ED2_timeoutput/Case2/run004/analy/"
first="tonzi-D-2003-07-01-000000-g01.h5"
last="tonzi-D-2003-08-01-000000-g01.h5"

#Dates from daily files
dates <- ExtractChronDates(dir,tscale="D")

#Dates from hourly files where each file has 24 hours in it
dates <- ExtractChronDates(dir,tscale="I", perfile=24)

#Dates from daily files between 07/01/2003 and 08/01/2003 (inclusive)
dates <- ExtractChronDates(dir,tscale="D", first=first, last=last)

}
