% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListAvailVars.R
\name{ListAvailVars}
\alias{ListAvailVars}
\title{List the ED2 variables that are available for extraction}
\usage{
ListAvailVars(direc, tscale, varsnippets)
}
\arguments{
\item{direc}{Directory where there are ED2 .h5 output files (character string)}

\item{tscale}{Time scale for relevant files (character vector, may have multiple entries)}

\item{varsnippets}{Text which the variable name should match (optional, character vector; without varsnippets all var names in tscale will print)}
}
\value{
Names of variables available by tscale
}
\description{
List the ED2 variables that are available for extraction
}
\examples{

direc<-"/n/moorcroftfs5/mjohnston/ED2_timeoutput/Case2/run004/analy/"

#Outputs all variable names in "D" or "I" files with the text "TEM" or "NEP"
ListAvailVars(direc=direc,tscale=c("D","I"),varsnippets=c("TEM","NEP"))

#Outputs all variable names in "D" files
#' ListAvailVars(direc=direc,tscale="D"))

}
