% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVars_SameAxis.R
\name{PlotVars_SameAxis}
\alias{PlotVars_SameAxis}
\title{Plot Output Variables on the Same Axis}
\usage{
PlotVars_SameAxis(dates, vars, names, units)
}
\arguments{
\item{dates}{'chron' dates, e.g. as extracted by ExtractChronDates function}

\item{vars}{list of (1 dimensional) variables that match dates - likely NOT raw output from an extraction (list of character strings)}

\item{names}{vector of names that will go on the legend (vector of character strings)}

\item{units}{Of the variables (string, optional)}
}
\value{
line plot with multiple data series
}
\description{
Given dates & 1-demensional variables with the same units, plot these on the same axis
}
\examples{

#Dates from daily files
PlotVars_SameAxis(chrondates,vars=list("L1","L2","L3"),names=c("leaftemp1","leaftemp2","leafttemp3"),units="K")
}
