% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVars_Stacked.R
\name{PlotVars_Stacked}
\alias{PlotVars_Stacked}
\title{Plot Variables on Separate Plots, align by date}
\usage{
PlotVars_Stacked(dates, vars, names)
}
\arguments{
\item{dates}{'chron' dates, e.g. as extracted by ExtractChronDates function}

\item{vars}{list of (1 dimensional) variables that match dates - likely NOT raw output from an extraction (list of character strings)}

\item{names}{vector of names that will go on the legend (vector of character strings)}
}
\value{
stacked line plots
}
\description{
Given dates & 1-demensional variables with the same units, plot these on the separate axes
}
\examples{

#Dates from daily files
PlotVars_Stacked(chrondates,vars=list("AGB","LEAF_TEMP","LAI"),names=c("biomass","leaftemp","lai"))
}
