% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLtoDataFrame.R
\name{XMLtoDataFrame}
\alias{XMLtoDataFrame}
\title{Extract PFT parameters from an ED2 XML}
\usage{
XMLtoDataFrame(xml)
}
\arguments{
\item{xml}{Full path to the XML file (character string)}
}
\value{
R Dataframe where column = PFT, row.names = parameter names and values are parameter values.
}
\description{
Extract PFT parameters from an ED2 XML
}
\examples{

xml<-"/mnt/odyssey/moorcroftfs5/mjohnston/gitruns/run006/Tonzi.xml"
df<-XMLtoDataFrame(xml)
#Note: warning "incomplete final line found on 'xmlfile.xml'" seems not to affect results.

}
